# Final stage of #3256 - Cleanup, annotation, and performance