… Terminal.Gui/MainLoop.cs
… Terminal.Gui/Terminal.Gui.csproj
… Terminal.Gui/Terminal.Gui.csproj.DotSettings
… Terminal.Gui/Text/Autocomplete/PopupAutocomplete.cs
… Terminal.Gui/Text/TextFormatter.cs
… Terminal.Gui/Types/Point.TemporaryOperators.cs
… Terminal.Gui/Types/Point.cs
… Terminal.Gui/Types/PointF.cs
… Terminal.Gui/Types/RectangleF.cs
… Terminal.Gui/Types/Size.TemporaryOperators.cs
… Terminal.Gui/Types/Size.cs
… Terminal.Gui/Types/SizeF.cs
… Terminal.Gui/View/Adornment/Adornment.cs
… Terminal.Gui/View/Adornment/Border.cs
… Terminal.Gui/View/Layout/ViewLayout.cs
… Terminal.Gui/View/View.cs
… Terminal.Gui/View/ViewDrawing.cs
… Terminal.Gui/View/ViewText.cs
… Terminal.Gui/Views/ColorPicker.cs
… Terminal.Gui/Views/GraphView/GraphView.cs
… Terminal.Gui/Views/HexView.cs
… Terminal.Gui/Views/Menu/Menu.cs
… Terminal.Gui/Views/MessageBox.cs
… Terminal.Gui/Views/RadioGroup.cs
… Terminal.Gui/Views/ScrollView.cs
… Terminal.Gui/Views/Slider.cs
… Terminal.Gui/Views/TableView/TableView.cs
… Terminal.Gui/Views/TextView.cs
… Terminal.Gui/Views/TileView.cs
… Terminal.Gui/Views/Toplevel.cs