This is the last of the changes in this set.

This one touches a few dozen files, but _most_ of the changes are just cleanup and don't affect compiled output. Primarily, those are removing redundant explicit type specifiers on constructor calls for Size, Rectangle, and Point.

I also fixed some test cases that were overlooked in earlier PRs that only affect tests run in Release mode (a couple more string cases that needed to be updated).

However, there _are_ some changes that _do_ affect compiled output, but **do not affect behavior**. These are performance-targeted changes such as reducing copies, using collection expressions, more efficient switches, and simpler operations in some hot paths. These resulted in tangible benefits. In crude testing, Debug builds gained an average of 13 seconds with stddev of 1.8s and release builds gained an average of 9 seconds with stddev of 1.1s vs the code prior to this change set (meaning everything up to and including #3269).

I also added a bunch of comments that you can see in the ReSharper TODO explorer (ctrl-alt-D by default) pointing out some issues and asking some questions. Or you can see them in the diffs, of course.

I had more I wanted to do, but I cut myself off since it was already going to touch quite a few files, for sake of our sanity. :)

## Proposed Changes/Todos

- [x] Clean up redundancies around creation of Size, Point, and Rectangle values
- [x] Add todo-style code comments
- [x] Performance enhancements
- [x] General cleanup in places I was already touching because of the above
- [x] Fixed remaining broken tests in Release build configuration

## Pull Request checklist:

- [x] I've named my PR in the form of "Fixes #issue. Terse description."
  - Sorta
- [x] My code follows the [style guidelines of Terminal.Gui](https://github.com/gui-cs/Terminal.Gui/blob/develop/.editorconfig) - if you use Visual Studio, hit `CTRL-K-D` to automatically reformat your files before committing.
- [x] My code follows the [Terminal.Gui library design guidelines](https://github.com/gui-cs/Terminal.Gui/blob/develop/CONTRIBUTING.md)
- [x] I ran `dotnet test` before commit
- [x] I have made corresponding changes to the API documentation (using `///` style comments)
- [x] My changes generate no new warnings
  - Many of the changes are actually explicitly to address existing warnings.
- [x] I have checked my code and corrected any poor grammar or misspellings
- [x] I conducted basic QA to assure all features are working


This one is also ready to go.

However, while it is cumulative with all the prior stages, I'd suggest it not be reviewed for merge until after #3269 is merged, to keep the critical changes separate from these, which are more discretionary.