@dodexahedron 2024-02-26
RE: The performance numbers in the OP...

Those numbers of course will be different on different hardware and whatnot. Those were from a 10th generation 6-core i7 (so 4 generations old at this point), on Windows, run via `dotnet test`.