@dodexahedron 2024-02-26
Had a stashed set of comments I meant to include.

Just added those.

That's it for this branch. :)